requirements.txt
requests
whisper
PyPDF2
python-dotenv
openai
torch
torchvision
torchaudio
gradio